* The OPTION_ macros are now integer(8). * The output format is now represented as a integer(8).

+++ src/td/pes_out.F90
@@ -76,7 +76,7 @@ contains
     character(len=*),  intent(in) :: file
     FLOAT,             intent(in) :: Lk(:,:)
     integer,           intent(in) :: ll(:)  
-    integer,           intent(in) :: how
+    integer(8),        intent(in) :: how
     type(simul_box_t), intent(in) :: sb 
     FLOAT, optional,   intent(in) :: pmesh(:,:,:,:)
